fix: corrections suite audit couches dao/controller/service
- ChienDao: corrige le type d'ID Long -> Integer pour matcher l'entite

- TypeSeanceDao: supprime import Seance inutilise

- RaceController: ajoute @JsonView(RaceView.class) sur les endpoints

  renvoyant une Race

- RaceController: supprime le no-op setNom(getNom()) dans updateRace

- RaceController: uniformise @PathVariable int -> Integer

- Race: retire @JsonView(RaceView.class) sur la collection chiens

  pour eviter la serialisation d'objets vides

feat(exception): introduit un package exception/ dedie

- Extrait SeanceNotFoundException dans son propre fichier (RuntimeException

  final, porte l'id manquant dans son message)

- Ajoute un @ExceptionHandler dedie dans GlobalExceptionInterceptor

  pour traduire l'exception en reponse HTTP 404

- SeanceService: supprime la classe interne et le throws

+++ src/main/java/edu/mns/cda/clubcaninbackend/dao/ChienDao.java
@@ -3,4 +3,4 @@ package edu.mns.cda.clubcaninbackend.dao;
 import edu.mns.cda.clubcaninbackend.model.Chien;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChienDao extends JpaRepository<Chien, Long> {}
+public interface ChienDao extends JpaRepository<Chien, Integer> {}
